+++ apps/router-demo/router-host-vue3-2100/rsbuild.config.ts
@@ -1,0 +1,37 @@
+import { defineConfig } from '@rsbuild/core';
+import { pluginVue } from '@rsbuild/plugin-vue';
+import { ModuleFederationPlugin } from '@module-federation/enhanced/rspack';
+import path from 'path';
+
+export default defineConfig({
+  source: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+      vue: path.resolve(__dirname, 'node_modules/vue'),
+      'vue-router': path.resolve(__dirname, 'node_modules/vue-router'),
+    },
+  },
+  server: {
+    port: 2100,
+  },
+  dev: {
+    // It is necessary to configure assetPrefix, and in the production environment, you need to configure output.assetPrefix
+    assetPrefix: 'http://localhost:2100',
+    writeToDisk: true,
+  },
+  tools: {
+    rspack: (config, { appendPlugins }) => {
+      delete config.optimization?.splitChunks;
+      appendPlugins([
+        new ModuleFederationPlugin({
+          name: 'host2',
+          remotes: {
+            remote1: 'remote1@http://localhost:2001/mf-manifest.json',
+          },
+        }),
+      ]);
+    },
+  },
+
+  plugins: [pluginVue()],
+});