+++ packages/bridge/bridge-shared/vite.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from 'vite';
+import vue from '@vitejs/plugin-vue';
+import path from 'path';
+import dts from 'vite-plugin-dts';
+import vueJsx from '@vitejs/plugin-vue-jsx';
+
+export default defineConfig({
+  plugins: [vue(), dts(), vueJsx()],
+  build: {
+    lib: {
+      entry: path.resolve(__dirname, 'src/index.ts'),
+      name: 'bridge-shared',
+      fileName: (format) => `index.${format}.js`,
+    },
+  },
+});