+++ packages/runtime/src/core.ts
@@ -105,7 +105,7 @@ export class FederationHost {
     fetch: new AsyncHook<
       [string, RequestInit],
       Promise<Response> | void | false
-    >('fetch'),
+    >(),
   });
 
   constructor(userOptions: UserOptions) {
@@ -204,8 +204,9 @@ export class FederationHost {
   initShareScopeMap(
     scopeName: string,
     shareScope: ShareScopeMap[string],
+    extraOptions: { hostShareScopeMap?: ShareScopeMap } = {},
   ): void {
-    this.sharedHandler.initShareScopeMap(scopeName, shareScope);
+    this.sharedHandler.initShareScopeMap(scopeName, shareScope, extraOptions);
   }
 
   private formatOptions(