+++ packages/sdk/src/dom.ts
@@ -64,14 +64,14 @@ export function createScript(info: {
         if (createScriptRes.timeout) timeout = createScriptRes.timeout;
       }
     }
-
     const attrs = info.attrs;
     if (attrs) {
       Object.keys(attrs).forEach((name) => {
         if (script) {
           if (name === 'async' || name === 'defer') {
             script[name] = attrs[name];
-          } else {
+            // Attributes that do not exist are considered overridden
+          } else if (!script.getAttribute(name)) {
             script.setAttribute(name, attrs[name]);
           }
         }
@@ -160,7 +160,7 @@ export function createLink(info: {
     const attrs = info.attrs;
     if (attrs) {
       Object.keys(attrs).forEach((name) => {
-        if (link) {
+        if (link && !link.getAttribute(name)) {
           link.setAttribute(name, attrs[name]);
         }
       });