+++ packages/webpack-bundler-runtime/src/initContainerEntry.ts
@@ -26,7 +26,9 @@ export function initContainerEntry(
     ...remoteEntryInitOptions,
   });
 
-  federationInstance.initShareScopeMap(name, shareScope);
+  federationInstance.initShareScopeMap(name, shareScope, {
+    hostShareScopeMap: remoteEntryInitOptions?.shareScopeMap || {},
+  });
 
   if (webpackRequire.federation.attachShareScopeMap) {
     webpackRequire.federation.attachShareScopeMap(webpackRequire);