+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,48 @@
 # [3.1.0](https://github.com/module-federation/core/compare/utils-3.0.2...utils-3.1.0) (2023-10-26)
 
+## 3.0.30
+
+### Patch Changes
+
+- Updated dependencies [32f26af]
+- Updated dependencies [32f26af]
+  - @module-federation/sdk@0.2.3
+
+## 3.0.29
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.2
+
+## 3.0.28
+
+### Patch Changes
+
+- Updated dependencies [88445e7]
+  - @module-federation/sdk@0.2.1
+
+## 3.0.27
+
+### Patch Changes
+
+- @module-federation/sdk@0.2.0
+
+## 3.0.26
+
+### Patch Changes
+
+- Updated dependencies [88900ad]
+  - @module-federation/sdk@0.1.21
+
+## 3.0.25
+
+### Patch Changes
+
+- Updated dependencies [685c607]
+- Updated dependencies [e8e0969]
+- Updated dependencies [349c381]
+  - @module-federation/sdk@0.1.20
+
 ## 3.0.24
 
 ### Patch Changes